refactor: 스토리북 일괄 수정
